Float returned in sixth.

+++ lab_01_06_05/main.c
@@ -2,14 +2,14 @@
 #include <math.h>
 #define EPS 1e-5
 
-double det(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4);
-int point_check(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4);
-int cramers_rule(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4, double det);
+float det(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4);
+int point_check(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4);
+int cramers_rule(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, float det);
 
 int main(void)
 {
-    double xp, yp, xq, yq;
-    double xr, yr, xs, ys;
+    float xp, yp, xq, yq;
+    float xr, yr, xs, ys;
     int rc1, rc2;
 
     printf("Введите координаты точек первого отрезка: ");
@@ -22,7 +22,7 @@ int main(void)
     if (rc2 != 4)
         return -100;
 
-    double determinant = det(xp, yp, xq, yq, xr, yr, xs, ys);
+    float determinant = det(xp, yp, xq, yq, xr, yr, xs, ys);
     if (fabs(determinant) <= EPS)
     {
         if (point_check(xp, yp, xq, yq, xr, yr, xs, ys))
@@ -40,20 +40,20 @@ int main(void)
     return 0;
 }
 
-double det(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4)
+float det(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4)
 {
     return (y4 - y3) * (x1 - x2) - (x4 - x3) * (y1 - y2);
 }
 
-int point_check(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4)
+int point_check(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4)
 {
     return (x1 * y2 - x2 * y1) * (x4 - x3) - (x3 * y4 - x4 * y3) * (x2 - x1) == 0 &&
         (x1 * y2 - x2 * y1) * (y4 - y3) - (x3 * y4 - x4 * y3) * (y2 - y1) == 0;
 }
 
-int cramers_rule(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4, double det)
+int cramers_rule(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, float det)
 {
-    double u_a, u_b, a, b;
+    float u_a, u_b, a, b;
     a = (x4 - x2) * (y4 - y3) - (x4 - x3) * (y4 - y2);
     b = (x1 - x2) * (y4 - y2) - (x4 - x2) * (y1 - y2);
     u_a = a / det;
